+++ packages/libs/hd-wallet/src/bip44/index.ts
@@ -1,4 +1,3 @@
-export * from './kadenaChangePassword';
 export * from './kadenaGenKeypairFromSeed';
 export * from './kadenaGetPublic';
 export * from './mnemonic';
