+++ packages/libs/hd-wallet/src/bip44/utils/sign.ts
@@ -2,18 +2,7 @@ import type { IUnsignedCommand } from '@kadena/client';
 import { sign } from '@kadena/cryptography-utils';
 
 import { HDKey } from 'ed25519-keygen/hdkey';
-
-/**
- * Convert a Uint8Array to a hexadecimal string.
- * @param {Uint8Array} uint8Array - The array to convert.
- * @returns {string} - Returns the hexadecimal representation of the input.
- */
-export const uint8ArrayToHex = (uint8Array: Uint8Array): string => {
-  if (uint8Array.length === 33 && uint8Array.at(0) === 0) {
-    uint8Array = uint8Array.slice(1);
-  }
-  return [...uint8Array].map((x) => x.toString(16).padStart(2, '0')).join('');
-};
+import { uint8ArrayToHex } from '../../utils/utils';
 
 /**
  * Derive a key pair using a seed and an index.
