+++ packages/libs/hd-wallet/src/utils/utils.ts
@@ -89,3 +89,15 @@ export function base64ToBuffer(base64: string): Buffer {
 
 export const HARDENED_OFFSET = 0x80000000;
 export const harden = (n: number) => HARDENED_OFFSET + n;
+
+/**
+ * Convert a Uint8Array to a hexadecimal string.
+ * @param {Uint8Array} uint8Array - The array to convert.
+ * @returns {string} - Returns the hexadecimal representation of the input.
+ */
+export const uint8ArrayToHex = (uint8Array: Uint8Array): string => {
+  if (uint8Array.length === 33 && uint8Array.at(0) === 0) {
+    uint8Array = uint8Array.slice(1);
+  }
+  return [...uint8Array].map((x) => x.toString(16).padStart(2, '0')).join('');
+};
